pyproject.toml
[build-system]
requires = [
  "flit_core @ git+https://github.com/martenlienen/flit.git@namespace-packages#subdirectory=flit_core",
  # Setuptools is just for setup.py and editable installs
  "setuptools"
]
build-backend = "flit_core.buildapi"

[project]
name = "hydra-experiment-sweeper"
version = "0.1.0"
description = "A hydra sweeper with configurable overrides for reproducible experiments"
readme = "README.md"
requires-python = ">= 3.8"
license = {file = "LICENSE"}
authors = [{name = "Marten Lienen", email = "marten.lienen@gmail.com"}]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python"
]
dependencies = [
  "hydra-core",
  "more-itertools"
]

[project.optional-dependencies]
test = ["pytest"]

[project.urls]
homepage = "https://github.com/martenlienen/hydra-experiment-sweeper"

[tool.flit.module]
name = "hydra_plugins"

[tool.flit.sdist]
exclude = ["tests/"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.pytest.ini_options]
testpaths = ["tests"]
